requirements.txt
asgiref==3.11.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
Faker==40.1.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
modeltranslation==0.25
pillow==12.1.0
psycopg==3.3.2
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
uvicorn==0.40.0
